Move the download resources function
